+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Change log
 
 - Refactored parameter definition support.
 - Updated `Krizalys\Onedrive\Proxy\FileProxy`.
+- Updated `Krizalys\Onedrive\Proxy\FolderProxy`.
 
 [2.4.2] - 2019-09-16
 --------------------